+++ src/test/java/com/softwareonpurpose/uinavigator/UiHostTest.java
@@ -23,7 +23,7 @@ public class UiHostTest extends TestBase {
     public void getInstance() {
         Class expected = UiHost.class;
         Class actual = UiHost.getInstance().getClass();
-        Assert.assertEquals(actual, expected, "The creation method getInstance returns an instantiated UiHost object");
+        Assert.assertEquals(actual, expected, "The creation method construct returns an instantiated UiHost object");
     }
 
     @Test(dependsOnMethods = "getInstance")
